fix(lint): restore clean cargo clippy --all-targets --all-features (#112)
* fix(lint): restore clean cargo clippy --all-targets --all-features

Fixes every warning that caused `cargo clippy --all-targets --all-features
-- -D warnings` to fail, bringing test/bench/doc targets to the same
strict lint posture as lib targets.

Changes per lint:
- needless_range_loop: use slice range iter in recursive_rules_test
- collapsible_if: collapse nested if-let guards in udf_test (let-chains)
- for_loops_over_single_element: unwrap single bench tuple in minigraf_bench
- unused_import under wasm: gate `use Indexes` with #[cfg(not(feature="wasm"))]
- to_string_in_format_args: drop redundant .to_string() on Uuid Display impls
- approx_constant (PI): replace 3.14/2.71 test values with 3.5/2.5
- cast_lossless / redundant cast: remove u64 as u64 in packed_pages test
- manual_arithmetic_check: use .saturating_sub(1) in persistent_facts test
- create_without_truncate: add .truncate(true) to OpenOptions in persistent_facts test
- clone_on_ref_ptr: use std::slice::from_ref instead of &[fact.clone()] in wal test
- useless_vec: replace vec![...] with array literals in functions tests
- doc_lazy_continuation: add blank /// lines after numbered list items in not_join_test

Closes #110


* style: apply cargo fmt to lint fixes


---------

+++ src/storage/packed_pages.rs
@@ -196,7 +196,7 @@ mod tests {
             Uuid::from_u128(n as u128),
             ":attr".to_string(),
             Value::Integer(n as i64),
-            n as u64,
+            n,
             n,
             0,
             VALID_TIME_FOREVER,
